Restore as int cast in Column.countNulls()
Groovy's static type checker infers count() return type as Number;
the cast is required for the method's int return type to compile.

+++ matrix-core/src/main/groovy/se/alipsa/matrix/core/Column.groovy
@@ -259,7 +259,7 @@ class Column extends ArrayList {
    * @return the count of null elements
    */
   int countNulls() {
-    count { it == null }
+    count { it == null } as int
   }
 
   List removeNulls() {
